refactor(core): decompose StreamCoordinator into focused modules
Break the 3654-line StreamCoordinator god object into 5 focused modules:
- PlaybackSessionStore: session CRUD with dual-indexed O(1) lookups
- VolumeRouter: sync-aware volume/mute routing
- SubscriptionArbiter: RC/GRC GENA subscription conflict resolution
- SyncGroupManager: sync lifecycle, stop orchestration, group restoration
- StreamCoordinator: thin orchestrator, stream CRUD, startup paths
