+++ packages/thaumic-core/src/services/mod.rs
@@ -6,10 +6,14 @@
 pub mod discovery_service;
 pub mod gena_event_processor;
 pub mod latency_monitor;
+pub mod playback_session_store;
 pub mod stream_coordinator;
+pub(crate) mod sync_group_manager;
 pub mod topology_monitor;
+pub mod volume_router;
 
 pub use discovery_service::DiscoveryService;
 pub use latency_monitor::LatencyMonitor;
-pub use stream_coordinator::{PlaybackResult, PlaybackSession, StreamCoordinator};
+pub use playback_session_store::{GroupRole, PlaybackResult, PlaybackSession};
+pub use stream_coordinator::StreamCoordinator;
 pub use topology_monitor::{TopologyMonitor, TopologyMonitorConfig};
